+++ src/roboto/version/autogen_version.py
@@ -1,4 +1,4 @@
-AUTOGEN_VERSION = "0.43.0"
+AUTOGEN_VERSION = "0.44.0"
 
 __all__= ("AUTOGEN_VERSION",)